Split template tab implementation into multiple files
