+++ plugins/woocommerce-beta-tester/src/product-editor-dev-tools/index.scss
@@ -218,6 +218,8 @@
 
 
 .woocommerce-product-editor-dev-tools-template-block {
+	margin-top: var(--woocommerce-product-editor-dev-tools-gap-small);
+
 	&__header {
 		font-weight: bold;
 	}
