+++ plugins/woocommerce-beta-tester/src/product-editor-dev-tools/template-tab-panel/index.tsx
@@ -7,72 +7,9 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
+import { BlockTemplateArray, BlockTemplate } from './types';
 import { TabPanel } from '../tab-panel';
-
-interface BlockTemplateAttributes {
-	[ key: string ]: unknown;
-	_templateBlockId: string;
-	_templateBlockOrder: number;
-}
-
-interface BlockTemplate {
-	0: string;
-	1: BlockTemplateAttributes;
-	2?: BlockTemplate[];
-}
-
-type BlockTemplateArray = BlockTemplate[];
-
-function BlockTemplateTreeItem( {
-	blockTemplate,
-	onSelect,
-}: {
-	blockTemplate: BlockTemplate;
-	onSelect: ( blockTemplateId: string ) => void;
-} ) {
-	const name = blockTemplate[ 0 ];
-	const attributes = blockTemplate[ 1 ];
-	const innerBlocks = blockTemplate[ 2 ];
-
-	const templateBlockId = attributes?._templateBlockId;
-	const templateBlockOrder = attributes?._templateBlockOrder;
-
-	function onClick( event: React.MouseEvent ) {
-		event.stopPropagation();
-
-		onSelect( templateBlockId );
-	}
-
-	return (
-		<div
-			className="woocommerce-product-editor-dev-tools-template-block"
-			onClick={ onClick }
-		>
-			<div className="woocommerce-product-editor-dev-tools-template-block__header">
-				{ templateBlockId }
-			</div>
-			<div className="woocommerce-product-editor-dev-tools-template-block__sub-header">
-				<span className="woocommerce-product-editor-dev-tools-template-block__name">
-					{ name }
-				</span>
-				<span className="woocommerce-product-editor-dev-tools-template-block__order">
-					{ templateBlockOrder }
-				</span>
-			</div>
-			{ innerBlocks && (
-				<div className="woocommerce-product-editor-dev-tools-template__inner-blocks">
-					{ innerBlocks.map( ( innerBlockTemplate, index ) => (
-						<BlockTemplateTreeItem
-							blockTemplate={ innerBlockTemplate }
-							key={ index }
-							onSelect={ onSelect }
-						/>
-					) ) }
-				</div>
-			) }
-		</div>
-	);
-}
+import { BlockTemplateTree } from './block-template-tree';
 
 export function TemplateTabPanel( {
 	isSelected,
@@ -120,15 +57,10 @@ export function TemplateTabPanel( {
 	return (
 		<TabPanel isSelected={ isSelected }>
 			<div className="woocommerce-product-editor-dev-tools-template">
-				<div className="woocommerce-product-editor-dev-tools-template-tree">
-					{ template.map( ( blockTemplate, index ) => (
-						<BlockTemplateTreeItem
-							blockTemplate={ blockTemplate }
-							key={ index }
-							onSelect={ onBlockTemplateSelect }
-						/>
-					) ) }
-				</div>
+				<BlockTemplateTree
+					template={ template }
+					onSelect={ onBlockTemplateSelect }
+				/>
 				<div className="woocommerce-product-editor-dev-tools-block-template-details">
 					{ JSON.stringify( selectedBlockTemplate, null, 4 ) }
 				</div>
