+++ plugins/woocommerce-beta-tester/src/product-editor-dev-tools/template-tab-panel/types.ts
@@ -1,0 +1,13 @@
+export interface BlockTemplateAttributes {
+	[ key: string ]: unknown;
+	_templateBlockId: string;
+	_templateBlockOrder: number;
+}
+
+export interface BlockTemplate {
+	0: string;
+	1: BlockTemplateAttributes;
+	2?: BlockTemplate[];
+}
+
+export type BlockTemplateArray = BlockTemplate[];
